# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Services à la Demande", "version": "0.1.0" },
  "paths": {
...
    "/proposals": {
      "post": {
        "tags": ["Proposals"],
        "summary": "Submit Proposal",
        "operationId": "submit_proposal_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProposalCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/proposals/request/{request_id}": {
      "get": {
        "tags": ["Proposals"],
        "summary": "List By Request",
        "operationId": "list_by_request_proposals_request__request_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Request Id" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "title": "Page" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/proposals/mine": {
      "get": {
        "tags": ["Proposals"],
        "summary": "My Proposals",
        "operationId": "my_proposals_proposals_mine_get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "title": "Page" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}/accept": {
      "post": {
        "tags": ["Proposals"],
        "summary": "Accept Proposal",
        "operationId": "accept_proposal_proposals__proposal_id__accept_post",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Proposal Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}/decline": {
      "post": {
        "tags": ["Proposals"],
        "summary": "Decline Proposal",
        "operationId": "decline_proposal_proposals__proposal_id__decline_post",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Proposal Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/proposals/{proposal_id}": {
      "delete": {
        "tags": ["Proposals"],
        "summary": "Delete Proposal",
        "operationId": "delete_proposal_proposals__proposal_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Proposal Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/messages/room": {
      "post": {
        "tags": ["Messages"],
        "summary": "Get Or Create Room",
        "description": "Résout le room_id pour une demande et un correspondant.\nLe résultat est déterministe : appeler deux fois avec les mêmes paramètres\nrenvoie toujours le même room_id.",
        "operationId": "get_or_create_room_messages_room_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RoomCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/messages/conversations": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get Conversations",
        "operationId": "get_conversations_messages_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/messages/{room_id}/history": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get History",
        "operationId": "get_history_messages__room_id__history_get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Room Id" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "title": "Page" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/messages/{room_id}": {
      "post": {
        "tags": ["Messages"],
        "summary": "Send Message",
        "operationId": "send_message_messages__room_id__post",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Room Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessageSend" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/messages/{room_id}/read": {
      "patch": {
        "tags": ["Messages"],
        "summary": "Mark Read",
        "operationId": "mark_read_messages__room_id__read_patch",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Room Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List Notifications",
        "operationId": "list_notifications_notifications_get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "title": "Page" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/notifications/unread-count": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Unread Count",
        "operationId": "unread_count_notifications_unread_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/notifications/read-all": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark All Read",
        "operationId": "mark_all_read_notifications_read_all_patch",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/notifications/{notif_id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark Read",
        "operationId": "mark_read_notifications__notif_id__read_patch",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Notif Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageSend": {
        "properties": {
          "content": { "type": "string", "title": "Content" },
          "media_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Media Url"
          },
          "request_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": ["content"],
        "title": "MessageSend"
      },
      "ProposalCreate": {
        "properties": {
          "request_id": { "type": "string", "title": "Request Id" },
          "message": { "type": "string", "title": "Message" },
          "price_estimate": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Price Estimate"
          }
        },
        "type": "object",
        "required": ["request_id", "message"],
        "title": "ProposalCreate"
      },
...
      "RoomCreate": {
        "properties": {
          "request_id": { "type": "string", "title": "Request Id" },
          "other_user_id": { "type": "string", "title": "Other User Id" }
        },
        "type": "object",
        "required": ["request_id", "other_user_id"],
        "title": "RoomCreate",
        "description": "Corps pour POST /messages/room — résout le room_id côté backend."
      },
      "RoomOut": {
        "properties": { "room_id": { "type": "string", "title": "Room Id" } },
        "type": "object",
        "required": ["room_id"],
        "title": "RoomOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/auth/login" } }
      }
    }
  }
}